Kinesis percentage fee implementation

+++ src/transactions/TransactionFrame.h
@@ -128,7 +128,7 @@ class TransactionFrame
         return mEnvelope.tx.sourceAccount;
     }
 
-    uint32_t getFee() const;
+    int64_t getFee() const;
 
     int64_t getMinFee(LedgerManager const& lm) const;
